requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
httpx==0.28.1
pydantic==2.11.7
python-dotenv==1.1.1
PyYAML==6.0.2
streamlit==1.48.1
requests==2.32.5
